Move date range utils into util file
