+++ src/providers/DateStoreProvider/internal/createScopedDateStore.tsx
@@ -1,11 +1,10 @@
 import { type PropsWithChildren, type ReactNode, useCallback, useMemo } from 'react'
 
-import type { DateRange, DateSelectionMode } from '@utils/date/dateRange'
+import { type DateRange, type DateSelectionMode, getEffectiveDateForMode, getEffectiveDateRangeForMode } from '@utils/date/dateRange'
 import { DatePreset, type SelectableDatePreset } from '@utils/date/dateRangePresets'
 import { createScopedStore } from '@utils/zustand/createScopedStore'
 import { useStoreWithDateSelected } from '@utils/zustand/useStoreWithDateSelected'
 import { buildDateStore, type MakeDateStoreOptions } from '@providers/DateStoreProvider/internal/buildDateStore'
-import { getEffectiveDateForMode, getEffectiveDateRangeForMode } from '@providers/DateStoreProvider/internal/dateStoreUtils'
 import { useBusinessActivationDateSafe, useDerivedInitialDateRange } from '@providers/DateStoreProvider/internal/useResolvedInitialRange'
 
 type DateStoreApi = ReturnType<typeof buildDateStore>
