+++ src/utils/date/dateRange.ts
@@ -1,4 +1,15 @@
-import { endOfDay, isEqual, max, min } from 'date-fns'
+import {
+  endOfDay,
+  endOfMonth,
+  endOfYear,
+  isEqual,
+  max,
+  min,
+  startOfMonth,
+  startOfYear,
+} from 'date-fns'
+
+import { unsafeAssertUnreachable } from '@utils/switch/assertUnreachable'
 
 export type DateSelectionMode = 'full' | 'month' | 'year'
 
@@ -14,4 +25,60 @@ export function clampToAfterActivationDate(date: Date | number, activationDate: 
 
 export function clampToPresentOrPast(date: Date | number, cutoff = endOfDay(new Date())) {
   return min([date, cutoff])
+}
+
+export function correctDateRange(range: DateRange): DateRange {
+  const { startDate, endDate } = range
+
+  if (startDate > endDate) {
+    return { startDate: endDate, endDate: startDate }
+  }
+
+  return range
+}
+
+type GetDateRangeOptions =
+  | { mode: 'full', startDate: Date, endDate: Date }
+  | { mode: Exclude<DateSelectionMode, 'full'>, startDate?: Date, endDate: Date }
+
+/**
+ * Shapes a range for a selection mode, clamping the end to the present:
+ * `month`/`year` snap to the period containing `endDate`; `full` keeps the
+ * explicit range.
+ */
+export function getDateRange(options: GetDateRangeOptions): DateRange {
+  const mode = options.mode
+  switch (mode) {
+    case 'month':
+      return {
+        startDate: startOfMonth(options.endDate),
+        endDate: clampToPresentOrPast(endOfMonth(options.endDate)),
+      }
+    case 'year':
+      return {
+        startDate: startOfYear(options.endDate),
+        endDate: clampToPresentOrPast(endOfYear(options.endDate)),
+      }
+    case 'full':
+      return {
+        startDate: options.startDate,
+        endDate: clampToPresentOrPast(endOfDay(options.endDate)),
+      }
+    default:
+      unsafeAssertUnreachable({
+        value: mode,
+        message: 'Invalid mode',
+      })
+  }
+}
+
+export function getEffectiveDateForMode(mode: DateSelectionMode, { date }: { date: Date }): { date: Date } {
+  return { date: getDateRange({ mode, startDate: date, endDate: date }).endDate }
+}
+
+export function getEffectiveDateRangeForMode(
+  mode: DateSelectionMode,
+  { startDate, endDate }: { startDate: Date, endDate: Date },
+): { startDate: Date, endDate: Date } {
+  return getDateRange({ mode, startDate, endDate })
 }
