+++ src/utils/date/dateRangePresets.ts
@@ -6,8 +6,7 @@ import {
   subYears,
 } from 'date-fns'
 
-import { clampToAfterActivationDate, clampToPresentOrPast, type DateRange } from '@utils/date/dateRange'
-import { correctDateRange } from '@providers/DateStoreProvider/internal/dateStoreUtils'
+import { clampToAfterActivationDate, clampToPresentOrPast, correctDateRange, type DateRange } from '@utils/date/dateRange'
 
 export enum Period {
   Month = 'Month',
